kata-deploy: support build shim-v2 for Arm CCA
Kernel 6.14 and above Ubuntu repo needs dependency: python3:native

Add cca-attester from coco-guest-component on Arm64 platform

After supporting the Arm CCA, it will rely on the kernel kvm.h headers to build the
runtime. The kernel-headers currently quite new with the traditional one, so that we
rely on build the kernel header first and then inject it to the shim-v2 build container.

Signed-off-by: Kevin Zhao <kevin.zhao@linaro.org>